Add project files

+++ compras/views.py
@@ -1,24 +1,27 @@
 from django.contrib.auth.decorators import login_required
-from django.http import HttpRequest
+from django.http import HttpRequest, HttpResponse
 from django.shortcuts import redirect, render
 
-from compras.models import Compra, CompraProduto
+from compras.models import Compra
 
 
 @login_required
 def visualizar_compras(request: HttpRequest):
-    total_compras = []
+    if not request.user.is_superuser:
+        return HttpResponse('Permissão insuficiente')
+    valores_totais = []
     compras = Compra.objects.all()
     for compra in compras:
         total = 0
-        compra_produtos = CompraProduto.objects.filter(compra=compra)
-        for compra_produto in compra_produtos:
+        for compra_produto in compra.compra_produtos.all():
             total += compra_produto.produto.preco * compra_produto.quantidade
-        total_compras.append(total)
-    compras = {k: v for k, v in zip(compras, total_compras)}
-    return render(request, 'compras/compras.html', {'compras': compras})
+        valores_totais.append(total)
+    compras_valores = {k: v for k, v in zip(compras, valores_totais)}
+    return render(request, 'compras/compras.html', {'compras_valores': compras_valores})
 
 @login_required
 def excluir_compra(request: HttpRequest, id):
+    if not request.user.is_superuser:
+        return HttpResponse('Permissão insuficiente')
     Compra.objects.get(id=id).delete()
     return redirect('visualizar_compras')
